+++ erp/imports/importer.py
@@ -1,86 +1,70 @@
-import logging
+from datetime import datetime
 
 from django.db import DataError, DatabaseError, transaction
 from django.db.transaction import TransactionManagementError
 
-from erp.imports.mapper.base import BaseRecordMapper
-from erp.imports.printer import outputPrintStrategy, outputVoidStrategy
-from erp.models import Activite, Accessibilite
+from erp.imports.mapper import SkippedRecord
 
-logger = logging.getLogger(__name__)
+from erp.models import Accessibilite
+
+ROOT_DATASETS_URL = "https://www.data.gouv.fr/fr/datasets/r"
 
 
 class Importer:
-    output = None
-    errors = []
-    imported = 0
-    skipped = 0
-
-    def __init__(self, mapper: BaseRecordMapper, is_scheduler=False) -> None:
-        self.output = outputVoidStrategy if is_scheduler else outputPrintStrategy
-        assert mapper.dataset_url is not None
+    def __init__(self, id, fetcher, mapper, activite=None, verbose=False, today=None):
+        self.id = id
+        self.fetcher = fetcher
         self.mapper = mapper
+        self.activite = activite
+        self.verbose = verbose
+        self.today = today if today is not None else datetime.today()
 
-    def job(self, verbose=False):
-        # reinitialize class instance properties, as it's long-lived
-        # accross imports in a scheduler context
-        self.imported = 0
-        self.skipped = 0
-        self.errors = []
+    def print_char(self, msg):
+        if self.verbose:
+            print(msg, end="", flush=True)
 
-        try:
-            for erp in self.do_import():
-                self.output("." if erp else "S", "", True)
-        except RuntimeError as err:
-            logger.error(err)
+    def process(self):
+        results = {
+            "imported": [],
+            "skipped": [],
+            "unpublished": [],
+            "errors": [],
+        }
 
-        if verbose and len(self.errors) > 0:
-            self.output("Erreurs rencontrées :")
-            for error in self.errors:
-                self.output(f"- {error}")
-
-        self.output("Opération effectuée:")
-        self.output(f"- {self.imported} ERP(s) importé(s)")
-        self.output(f"- {self.skipped} écarté(s)")
-
-        return self.imported, self.skipped, self.errors
-
-    def do_import(self):
-        try:
-            data = self.mapper.fetch_data()
-        except RuntimeError as err:
-            raise RuntimeError("Error fetching data: ", str(err))
-
-        return self._process_data(data)
-
-    def _process_data(self, records):
-        activite = Activite.objects.filter(slug=self.mapper.activite).first()
-        if not activite:
-            raise RuntimeError(f"L'activité {activite} n'existe pas.")
-
-        for record in records:
+        for record in self.fetcher.fetch(f"{ROOT_DATASETS_URL}/{self.id}"):
             try:
+                (erp, unpublish_reason) = self.mapper(
+                    record, self.activite, self.today
+                ).process()
+                if not erp:
+                    self.log_char("X")
+                    continue
                 with transaction.atomic():
-                    erp = self.mapper.process(record, activite)
-                    self._save_erp(erp)
-                self.imported += 1
-                yield erp
+                    if unpublish_reason:
+                        erp.published = False
+                    erp.save()
+
+                    # Attach an Accessibilite to newly created Erps
+                    if not erp.has_accessibilite():
+                        accessibilite = Accessibilite(erp=erp)
+                        accessibilite.save()
+                    else:
+                        erp.accessibilite.save()
+
+                    if unpublish_reason:
+                        self.print_char("U")
+                        results["unpublished"].append(f"{str(erp)}: {unpublish_reason}")
+                    else:
+                        self.print_char(".")
+                        results["imported"].append(str(erp))
+            except SkippedRecord as skipped_reason:
+                self.print_char("S")
+                results["skipped"].append(f"{str(erp)}: {skipped_reason}")
             except RuntimeError as err:
-                self.errors.append(str(err))
-                self.skipped += 1
-                yield None
+                self.print_char("E")
+                results["errors"].append(f"{str(erp)}: {str(err)}")
+            except (TransactionManagementError, DataError, DatabaseError) as err:
+                self.print_char("E")
+                results["errors"].append(f"{str(erp)}: {str(err)}")
 
-    def _save_erp(self, erp):
-        try:
-            # Save erp instance
-            erp.published = True
-            erp.save()
-
-            # Attach an Accessibilite to newly created Erps
-            if not erp.has_accessibilite():
-                accessibilite = Accessibilite(erp=erp)
-                accessibilite.save()
-            else:
-                erp.accessibilite.save()
-        except (TransactionManagementError, DataError, DatabaseError) as err:
-            raise RuntimeError(f"Erreur à l'enregistrement des données: {err}") from err
+        return results
