--- erp/imports/printer.py
@@ -1,6 +1,0 @@
-def outputPrintStrategy(text, end="\n", flush=False):
-    print(text, end=end, flush=flush)
-
-
-def outputVoidStrategy(text, end="\n", flush=False):
-    pass
